requirements.txt
Django~=2.0.6
DjangoRestFramework~=3.9.0
oauth2client~=3.0.0
jsonpickle~=1.0